cart becomes empty without placing an order issue fixed

+++ app/scripts/controllers/goods.js
@@ -74,19 +74,21 @@
             priceNQT: shoppingCartGoods.priceNQT,
             deliveryDeadlineTimestamp: String(nxt.util.convertToEpochTimestamp(Date.now()) + 60 * 60 * 168)
           }
-          plugins.get('transaction').get('dgsPurchase').execute($scope.id_rs, order_args).then(function() {
-            $http({
-              url: shoppingCartGoods.callback,
-              data: shoppingCartGoods,
-              method: 'POST'
-            }).success(function(data) {
-              console.log(data);
-            }).error(function(err) {
-              console.log(err);
-            })
-            shoppingCartService.removeItem(0);
-            shoppingCart.splice(0, 1);
-            processCart(shoppingCart);
+          plugins.get('transaction').get('dgsPurchase').execute($scope.id_rs, order_args).then(function(data) {
+            if(data) {  
+              $http({
+                url: shoppingCartGoods.callback,
+                data: shoppingCartGoods,
+                method: 'POST'
+              }).success(function(data) {
+                console.log(data);
+              }).error(function(err) {
+                console.log(err);
+              })
+              shoppingCartService.removeItem(0);
+              shoppingCart.splice(0, 1);
+              processCart(shoppingCart);
+            }
           });
         }
       }
